Modify the Vagrantfile a little bit for bootstrap.
The bootstrap virtual machine can perform some Chef provisioning. In the
prior implementation we do a few funky things with VB configuration that
I carried over. Not completely sure its necessary (e.g. defaults) but
left it in here just in case.

It looks like disks are attached to the virtual machine using the CLI
from the vbox_create.sh script. It'd be great if we can encompass that
in the Vagrantfile. I'll investigate that a little bit.